requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
pkg_resources==0.0.0
pycparser==2.21
pycrypto==2.6.1
pyOpenSSL==22.0.0
qrcode==7.3.1
simple-crypt==4.1.7
Werkzeug==2.0.2
